remove Components
